Changes - Updates
Updated to the last version of Unity and because of that had to change lots of things that were giving errors.

Added:
Spikes to the game.
Added a coroutine to the mana damage so it will look a bit better.
Added Death animations.
Added Sword attacks sounds.
Added functionality on the skill effects (Stun, Regen, Heal, Dps).
Skills Damage/heal increases by the amount of int.
Changed how the game windows rescale so it works better on each resolution (still funky).
Updated from text to Tesh mesh pro so had to change a few things.
- Finished the first floor.
- Added many weapons and armors
- Improved char menu, can now see stats there when you equip things.
- Added arcana damage
- Added tutorials for the User
- Can now load game from the Main menu

-Fixed
- Loot function properly.
- Enemies no longer attack you while you move or while he is moving.
- Slimes collider now change to trigger after death.
- Improved player attack and added CD.
- Fixed a bug in which enemies could attack each other.
- Fixed a bug in which the player could attack while death.
- Player could attack and still move after it without finishing animation.
- Changed enemy canvas.
- Hand now updates properly, same with slots in inventory.
- Killing enemies with Dps wouldn't change the enemy state to death.
- Fixed how bags swap and unequipping gear.
- Player was able to attack while casting
- Mana cost works properly and updates fine.
- Fixed some visual bugs on the stats and the skills extra damage value
- Saved games weren't showing up properly